Define fully exit fullscreen in terms of exitFullscreen({ fully: true })
The previous definition of fully exit fullscreen could unfullscreen
elements in top layer synchronously. If those elements were iframes,
the contentDocument would get handled in a following animation frame
task, leaving everything in an odd state in the interim.

Exposing the concept of fully exiting to scripts makes it easier to
test, although this is not central to the change itself.